actionlist: provide means of attaching shortcuts to windows
Prior, only shortcuts for main menus worked. With this, it's possible to
attach context menu shortcuts or toolbar shortcuts to arbitrary window
objects.

This isn't super elegant -- it'd be nice were this done automatically.
But even in the automatic case, it's possible that folks might want to
attach shortcuts to additional windows beyond the one that has the
toolbar or context menu.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>