Added WindowGroup for management of state with thread affinity
Each window now has a WindowGroup assigned to it as part of its InitWindow
process. All windows created on a common thread will share the same
window group.

Widgets now access their ToolTip controls through their window group.
The tool tip control is created on first use and is destroyed when the
group is destroyed.

Window groups are destroyed once all windows that refer to the group
are destroyed.

Window group data should always be accessed from the same thread. Despite
this, the data within each group is protected by an atomic counter and
a mutex as a safety measure. This is subject to change in the future.

This is an alternative solution for https://github.com/lxn/walk/issues/601