Upgrade icon management to reestablish examples/imageicon
SetIcon method of FormBase was using a big icon DPI calculation that was
wrong: on 192DPI it calculated 288 instead of 384. The calculation was
simplified: for the big icon, we want 2x the DPI of the small icon.
Usually.

Furthermore, NewIconFromImage(), NewIconFromBitmap() and
NewIconFromHICON() were extended to require dpi. This allows correct
initialization of the icon. NewIconFromImageWithSize() was upgraded to
calculate DPI from image sizes.

Icon's drawStretched method was upgraded to fallback to the maximum DPI
available icon. examples/imageicon creates the 16x16 icon dynamically.
When 32x32 icon is requested, handleForDPI() cannot load 2xDPI variant
from anywhere as the icon is not resource or file based. This leads to
the chicken&egg problem: IconCache is using drawStretched() to
create other DPI variants of the icon when that DPI variant is not
available yet.

Signed-off-by: Simon Rozman <simon@rozman.si>