Icon: Use 96dpi for icon load-check
The checkNewIcon() is used to test if the icon is loadable. Before, it
obtained screen DPI to load the appropriate icon size assuming it will
be used on the same DPI canvas or window. To avoid screen DPI detection,
this commit drops this assumption and just picks the hard-coded 96dpi
for the load-check.

Signed-off-by: Simon Rozman <simon@rozman.si>