notifyicon: recreate when explorer.exe crashes
According to [1], we receive a TaskbarCreated message after Windows
recreates the taskbar, which can happen if explorer.exe crashes or even
if the DPI changes. When this happens, we're supposed to create a new
notify icon; otherwise the previous one is nowhere to be found. This
commit listens for that message and recreates the state accordingly.

[1] https://docs.microsoft.com/en-us/windows/win32/shell/taskbar?redirectedfrom=MSDN#taskbar-creation-notification

Reported-by: Riccardo Paolo Bestetti <pbl@bestov.io>
Diagnosed-by: Simon Rozman <simon@rozman.si>